backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-mail==0.9.1
flask-bcrypt==1.0.1
PyJWT==2.8.0
Flask-PyMongo==2.3.0
dnspython==2.6.1
pandas==2.1.1
numpy==1.26.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
